Remove repetitious steps from acceptance cluster
Install and configure slurm through ansible for the acceptance cluster
too.  There is no point in me doing this each time I run acceptance for
a release.  It just takes time and is error prone taking more time and
I'm not performing acceptance of slurm.

Unfortunately, this does involve installing the openflight repo, but I
think that is an acceptable compromise.